requirements.txt
aiocontextvars==0.2.2
appdirs==1.4.4
APScheduler==3.2.0
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextvars==2.4
cryptography==2.1.4
cssselect==1.1.0
dataclasses==0.8
distro-info===0.18ubuntu0.18.04.1
fake-useragent==0.1.11
flake8==3.8.4
Flask==1.0
gunicorn==19.9.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.7
immutables==0.15
importlib-metadata==3.7.2
incremental==16.10.1
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
loguru==0.5.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
netifaces==0.10.4
PAM==0.4.2
parsel==1.6.0
pathspec==0.8.1
Protego==0.1.16
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.6.0
pycrypto==2.6.1
PyDispatcher==2.0.5
PyExecJS==1.5.1
pyflakes==2.2.0
pygobject==3.26.1
PyMySQL==1.0.2
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
queuelib==1.5.0
redis==3.5.3
regex==2020.11.13
requests==2.20.0
requests-unixsocket==0.1.5
Scrapy==2.4.1
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
ssh-import-id==5.7
systemd-python==234
toml==0.10.2
Twisted==17.9.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.3
w3lib==1.22.0
Werkzeug==0.15.3
zipp==3.4.1
zope.interface==4.3.2
